Module neure::neu

Structs

Alphabetic

Alphanumeric

And

Return true if the value matches both L and R.

Ascii

AsciiAlphabetic

AsciiAlphanumeric

AsciiControl

AsciiDigit

AsciiGraphic

AsciiHexDigit

AsciiLowercase

AsciiPunctuation

AsciiUppercase

AsciiWhiteSpace

CRange

Control

Digit

Equal

False

Lowercase

MayUnit

NeureOne

Repeat the unit U one time.

NeureOneMore

Repeat the unit U one or more times.

NeureRepeat

Repeat the match unit U at least M times and at most N times.

NeureRepeatRange

Repeat the unit U and the number of matches meet the given range.

NeureThen

Construct a regex that matches unit L and then unit R.

NeureZeroMore

Repeat the unit U zero or more times.

NeureZeroOne

Repeat the unit U zero or one time.

Not

Return true if the given value not matches U.

NullCond

Numeric

Or

Return true if the value matches L or R.

RegexCond

True

Uppercase

WhiteSpace

Wild

Traits

Condition

Neu

Neu2Re

NeuCond

NeuOp

Functions

alphabetic

Reference is_alphabetic.

alphanumeric

Reference is_alphanumeric.

and

Return true if the value matches both L and R.

any

Always return true.

ascii

Reference is_ascii or is_ascii.

ascii_alphabetic

Reference is_ascii_alphabetic or is_ascii_alphabetic.

ascii_alphanumeric

Reference is_ascii_alphanumeric or is_ascii_alphanumeric.

ascii_control

Reference is_ascii_control or is_ascii_control.

ascii_digit

Reference is_ascii_digit or is_ascii_digit.

ascii_graphic

Reference is_ascii_graphic or is_ascii_graphic.

ascii_hexdigit

Reference is_ascii_hexdigit or is_ascii_hexdigit.

ascii_lowercase

Reference is_ascii_lowercase or is_ascii_lowercase.

ascii_punctuation

Reference is_ascii_punctuation or is_ascii_punctuation.

ascii_uppercase

Reference is_ascii_uppercase or is_ascii_uppercase.

ascii_whitespace

Reference is_ascii_whitespace or is_ascii_whitespace.

control

Reference is_control.

digit

Reference is_digit.

equal

Match a character equal to given value.

lowercase

Reference is_lowercase.

may

Match if and remember the result, then match unit

may_count

none

Always return false.

not

Return true if the given value not matches U.

numeric

Reference is_numeric.

or

Return true if the value matches L or R.

range

Match a character within the specified range.

re_cond

Create a condition using in Condition base on regex.

uppercase

Reference is_uppercase.

whitespace

Reference is_whitespace.

wild

Match all the characters except \n.