Module neure::re

Re-exports

pub use self::ctor::branch;

pub use self::ctor::Array;

pub use self::ctor::ConstructOp;

pub use self::ctor::Ctor;

pub use self::ctor::DynamicArcCtor;

pub use self::ctor::DynamicBoxedCtor;

pub use self::ctor::DynamicBoxedCtorSync;

pub use self::ctor::DynamicCreateCtorThen;

pub use self::ctor::DynamicCreateCtorThenHelper;

pub use self::ctor::DynamicRcCtor;

pub use self::ctor::PairArray;

pub use self::ctor::PairSlice;

pub use self::ctor::PairVector;

pub use self::ctor::Slice;

pub use self::ctor::Vector;

pub use self::regex::AnchorEnd;

pub use self::regex::AnchorStart;

pub use self::regex::BoxedRegex;

pub use self::regex::Consume;

pub use self::regex::ConsumeAll;

pub use self::regex::DynamicArcRegex;

pub use self::regex::DynamicBoxedRegex;

pub use self::regex::DynamicCreateRegexThenHelper;

pub use self::regex::DynamicRcRegex;

pub use self::regex::LitSlice;

pub use self::regex::LitString;

pub use self::regex::RegexNot;

Modules

ctor

regex

Structs

NullRegex

Pass

RecParser

RecParserSync

WrappedTy

Traits

ConstructIntoOp

Extract

Handler

RecursiveParser

RecursiveParserSync

Regex

RegexIntoOp

Wrapped

Functions

array

consume

Consume given length datas.

consume_all

Consume all the left datas.

count

Match the given Neu M ..= N times.

count_if

Match the given Neu M ..= N times.

end

Match the end position of data.

lit_slice

Match given data.

not

Return a regex that reverses the result of re. It will return zero-length Span when matches.

null

Match nothing, simple return R::from(_, (0, 0)).

one

Match one item.

one_more

Match at least one item.

pair_array

pair_slice

pair_vector

Iterate over the vector and match the regex against the Context. It will return the value of first pair that matches.

rec_parser

For use in recursive parsers, it will construct a type that implements Ctor<'_, C, M, O> from a closure(Fn(&mut C) -> Result<O, Error>).

rec_parser_sync

rec_parser_with

Example

rec_parser_with_sync

slice

start

Match the start position of data.

string

Match given string.

vector

Iterate over the vector and match the regex against the Context. It will return the result of first regex that matches.

zero_more

Match at least zero item.

zero_one

Match zero or one item.

Type Aliases

RecursiveCtor

RecursiveCtorSync

RecursiveCtorWith

RecursiveCtorWithSync