Module neure::re::ctor

Structs

Array

Iterate over the array and match the regex against the Context.

BoxedCtor

Collect

Repeatedly match the regex P at least min times.

DynamicArcCtor

DynamicBoxedCtor

DynamicBoxedCtorSync

DynamicCreateCtorThen

Like Then, but create with .dyn_then_ctor of regex.

DynamicRcCtor

IfRegex

Construct a branch struct base on the test I(Fn(&C) -> Result<bool, Error>).

IfThen

First try to match P. If it succeeds, then try to match I. If it succeeds, then try to match T.

LongestTokenMatch

Match L and R, return the longest match result.

Map

Map the result to another type.

OptionPat

Match P and return the result wrapped by Option, ignoring the error.

Or

First try to match L, if it fails, then try to match R.

Pad

First try to match P. If the match succeeds, then try to match T.

Padded

First try to match T. If it succeeds, try to match P.

PairArray

Iterate over the array and match the regex against the Context.

PairSlice

Iterate over the slice and match the regex against the Context.

PairVector

Iterate over the vector and match the regex against the Context.

Pattern

Call .try_mat to match regex P.

Quote

First try to match L. If it is succeeds, then try to match P. If it is succeeds, then try to match R.

Repeat

Repeatedly match regex P, and the number of matches must meet the given range.

SepCollect

Match regex P as many times as possible, with S as the delimiter.

SepOnce

Match L and R separated by S.

Separate

Match regex P as many times as possible, with S as the delimiter.

Slice

Iterate over the array and match the regex against the Context.

Then

First try to match P. If it succeeds, then try to match T.

Vector

Iterate over the vector and match the regex against the Context.

Traits

ConstructOp

Ctor

DynamicCreateCtorThenHelper

Functions

branch