Module neure::re::ctor

Structs

BoxedCtor

Collect

Repeat the regex P and then collect into another type V.

DynamicCreateCtorThen

DynamicCtor

IfRegex

Construct a branch struct base on the test I(Fn(&C) -> Result<bool, Error>).

IfThen

Match P and then match T.

LongestTokenMatch

Match L and R, return the longest match result.

Map

Map the result to another type.

OptionPat

Or

Match L or R.

PadUnit

Match P and T which padded by P.

PaddedUnit

Match P which padded by T.

PairVector

Pattern

Ignore the interal struct of regex P, convert it to a single pattern.

Quote

Match regex P quoted by L and R.

Repeat

Match P repeatedly.

SepCollect

Match regex P as many times as possible, with S as the delimiter.

SepOnce

Match L and R separated by S.

Separate

Match regex P as many times as possible, with S as the delimiter.

Then

Match P and then match T.

Vector

Traits

BoxedCtorHelper

ConstructOp

Ctor

DynamicCreateCtorThenHelper

DynamicCtorHelper

Functions

branch

into_boxed_ctor

Construct a boxed Ctor that convert from M to O.

into_dyn_ctor

For use in recursive parsers, it will construct a type that implements Ctor<'_, C, M, O> from a closure(Fn(&mut C) -> Result<O, Error>).

rec_parser

rec_parser_sync

Type Aliases

DynamicCtorHandler

RecursiveCtor

RecursiveCtorSync